+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [v0.1.1] - 2025-05-12
+
+### Added
+- Added PublishBlogPostStep to handle publishing to GitHub using ModelContextProtocolGitHub integration for blog post publishing using ModelContextProtocol
+- Added UI components to allow users to publish blog posts to GitHub
+
+
 ## [v0.1.0] - 2025-04-23
 
 ### Added